Cleanup and dropping BarStat
BarStat was not going to provide much value and it was very complex
getting it wired up nicely with styled components. Since it was very
simple to begin with, just having a Bar.Wrapper and a Bar.Fill was
sufficient and not very complicated. It also allows for more
customization in individual use cases.