Solar System: Bar (and a few more adjustments) (#38)
* Adds a withDefaulTheme HOC for other components to use

* Update Paper to use README and splits styled component

* Wrap Badge examples in Paper components

* Updates BarStat and integrate into storybook

* A few more updates

* Cleanup and dropping BarStat

BarStat was not going to provide much value and it was very complex
getting it wired up nicely with styled components. Since it was very
simple to begin with, just having a Bar.Wrapper and a Bar.Fill was
sufficient and not very complicated. It also allows for more
customization in individual use cases.

* {geyset => geyser} typo

* Adds knobs for bg/fill in the Bar Overview story

* Use { Bar, Fill } import instead of redundant name

* Two final story adjustments

* Static overflow property and better height definitions