fix: validate image asynchronously by HTMLImageElement instead of XHR (#146)
* fix: validate image asynchronously by HTMLImageElement instead of XHR

Use an HTMLImageElement instance to validate image URLs instead of
checking the response of a HEAD request. The problem with that approach
is that if the servers don't allow HEAD requests, no images will be
loaded even though they are valid URLs. This solution also performs this
validation process asynchronously so no more blocking the main thread by
performing XHR requests.

Fixes #145

* refactor: rename loadedImageUrl state to loaded

* feat: add test case to check for race conditions

* feat: set HTTP headers if they are available

* build: bump version to v1.2.3